Delete comment authorization implemented

+++ src/main/java/com/example/BloggingApplication/controller/CommentController.java
@@ -5,6 +5,8 @@ import com.example.BloggingApplication.dto.ApiResponse;
 import com.example.BloggingApplication.dto.CreateComment;
 import com.example.BloggingApplication.dto.ResponseCommentDTO;
 import com.example.BloggingApplication.dto.UpdateCommentDTO;
+import com.example.BloggingApplication.exception.AuthorizationException;
+import com.example.BloggingApplication.service.AuthorizationService;
 import com.example.BloggingApplication.service.CommentService;
 import com.example.BloggingApplication.util.ResponseUtil;
 import jakarta.servlet.http.HttpServletRequest;
@@ -19,15 +21,20 @@ import java.util.List;
 @RequestMapping("api/comment")
 public class CommentController {
     private CommentService commentService;
-
+    private AuthorizationService authorizationService;
     @Autowired
-    public CommentController(CommentService commentService) {
+    public CommentController(CommentService commentService, AuthorizationService authorizationService) {
         this.commentService = commentService;
+        this.authorizationService = authorizationService;
     }
 
     @PostMapping("/")
     public ResponseEntity<ApiResponse<ResponseCommentDTO>> addComment(@Valid  @RequestBody CreateComment comment, HttpServletRequest request)
     {
+        if(!authorizationService.checkIfAuthorizeToComment(comment.getCommentatorId(),request))
+        {
+            throw new AuthorizationException("CommentatorId is not the same as the requestedId");
+        }
         return ResponseEntity.ok(ResponseUtil.success(commentService.addComment(comment), "comment added successfully", request.getRequestURI()));
     }
 
@@ -40,12 +47,20 @@ public class CommentController {
     @PutMapping("/")
     public ResponseEntity<ApiResponse<ResponseCommentDTO>> updateCommentByCommentId(@Valid @RequestBody UpdateCommentDTO updateCommentDTO, HttpServletRequest request)
     {
+        if(!authorizationService.checkIfAuthorizeToUpdateComment(updateCommentDTO.getCommentId(),request))
+        {
+            throw new AuthorizationException("Do not have authorization to update comment");
+        }
         return ResponseEntity.ok(ResponseUtil.success(commentService.updateComment(updateCommentDTO),String.format("Comment with commentId %d successfully updated",updateCommentDTO.getCommentId()), request.getRequestURI()));
     }
 
     @DeleteMapping("/{commentId}")
     public ResponseEntity<ApiResponse<Integer>>deleteCommentByCommentId(@Valid @PathVariable int commentId, HttpServletRequest request)
     {
+        if(!authorizationService.checkIfAuthorizeToDeleteComment(commentId,request))
+        {
+            throw new AuthorizationException("Does not have authorization to delete comment");
+        }
         return ResponseEntity.ok(ResponseUtil.success(commentService.deleteCommentByCommentId(commentId),String.format("Comment with commentId %d deleted",commentId), request.getRequestURI()));
     }
 }
